refactor cli
